#468 students count: develop: add semester as parameter; refactor: mapping of SelectiveCourseWithStudentsCountDTO

+++ api/src/main/java/ua/edu/chdtu/deanoffice/api/course/selective/dto/SelectiveCourseWithStudentsCountDTO.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class SelectiveCourseWithStudentsCountDTO extends SelectiveCourseDTO {
-    private Long studentsCount;
+    private int studentsCount;
 }
